requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
distlib==0.3.6
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-filter==23.1
django-timezone-field==5.0
djangorestframework==3.14.0
filelock==3.12.0
identify==2.5.23
kombu==5.2.4
Markdown==3.4.3
nodeenv==1.7.0
platformdirs==3.4.0
pre-commit==3.2.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.2
PyYAML==6.0
redis==4.5.3
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
vine==5.0.0
virtualenv==20.22.0
wcwidth==0.2.6
